backend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Deep Learning
torch>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Environment Variables
python-dotenv>=1.0.0
